feat: делит JS на модули
