+++ js/main.js
@@ -1,41 +1,3 @@
-const getRandomInteger = (min, max) => {
-  const fraction = Math.random() * (max - min + 1) + min;
-  return Math.floor(fraction);
-};
-
-const getRandomItem = (items) => {
-  const lastIndex = items.length - 1;
-  const index = getRandomInteger(0, lastIndex);
-  return items[index];
-};
-
-const createCommentsData = (itemCount) => {
-  const messages = [
-    'Всё отлично!',
-    'В целом всё неплохо. Но не всё.',
-    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-  ];
-  const names = ['Сергей', 'Ольга', 'Анна', 'Владимир', 'Николай', 'Артём', 'Виктор', 'Наталья', 'Надежда'];
-  return new Array(itemCount).fill(1).map((start, index) => ({
-    id: start + index,
-    avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-    message: getRandomItem(messages),
-    name: getRandomItem(names),
-  }));
-};
-
-const createPicturesData = (itemCount = 25) => {
-  const descriptions = ['Описание 1', 'Описание 2', 'Описание 3'];
-  return new Array(itemCount).fill(1).map((start, index) => ({
-    id: start + index,
-    url: `photos/${start + index}.jpg`,
-    description: getRandomItem(descriptions),
-    likes: getRandomInteger(15, 200),
-    comments: createCommentsData(getRandomInteger(0, 30)),
-  }));
-};
+import {createPicturesData} from './data.js';
 
 void(createPicturesData);
